Move wifi driver one module up
