+++ src/board/hardware/v1.rs
@@ -20,7 +20,7 @@ use crate::{
         },
         startup::WIFI_DRIVER,
         utils::{DummyOutputPin, SpiDeviceWrapper},
-        wifi::driver::WifiDriver,
+        wifi::WifiDriver,
         *,
     },
     heap::init_heap,
